Add ITupleSink interface and refactor code
