+++ core/dataflow/src/main/java/edu/uci/ics/texera/dataflow/sink/VisualizationOperator.java
@@ -1,5 +1,6 @@
 package edu.uci.ics.texera.dataflow.sink;
 
+import edu.uci.ics.texera.api.dataflow.IOperator;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -13,9 +14,11 @@ import edu.uci.ics.texera.api.tuple.Tuple;
  * Base class for visualization operators
  * Author: Mingji Han
  */
-public abstract class VisualizationOperator extends AbstractTupleSink {
+public abstract class VisualizationOperator implements ITupleSink {
 
-
+    protected IOperator inputOperator;
+    protected int cursor = CLOSED;
+    protected Schema outputSchema;
     protected List<Tuple> result = new ArrayList<>();
     protected final String type;
 
@@ -23,6 +26,25 @@ public abstract class VisualizationOperator extends AbstractTupleSink {
         this.type = type;
     }
 
+    @Override
+    public void setInputOperator(IOperator inputOperator) {
+        if (cursor != CLOSED) {
+            throw new TexeraException(ErrorMessages.INPUT_OPERATOR_CHANGED_AFTER_OPEN);
+        }
+        this.inputOperator = inputOperator;
+    }
+
+    @Override
+    public void close() throws TexeraException {
+        if (cursor == CLOSED) {
+            return ;
+        }
+
+        if (inputOperator != null)
+            inputOperator.close();
+
+        cursor = CLOSED;
+    }
 
 
     @Override
@@ -35,13 +57,16 @@ public abstract class VisualizationOperator extends AbstractTupleSink {
         return outputSchema;
     }
 
-
+    @Override
     public List<Tuple> collectAllTuples() {
         processTuples();
         return result;
     }
 
-
+    @Override
+    public Schema transformToOutputSchema(Schema... inputSchema) {
+        throw new TexeraException(ErrorMessages.INVALID_OUTPUT_SCHEMA_FOR_SINK);
+    }
 
     public String getChartType() {
         return type;
